modules/AWSMQTTModule/requirements.txt
pytz==2019.3
termcolor
paho-mqtt
